requirements.txt
Django                  == 5.0.4
djangorestframework     == 3.15.1
django-enumfield        == 3.1
requests                == 2.31.0
python-dotenv           == 1.0.0
psycopg2                == 2.9.7
psycopg2-binary         == 2.9.7
uwsgi